CoursierDownloader: split factory, move to own pkg
